data/list.h: Make LIST and LIST_NODE distinct types.
We used to have LIST_NODE and LIST as synonyms (typedef), because our
implementation uses the LIST as a (special) node to avoid branches in
insert and remove operations.

However it may cause problems because C compiler cannot distinguish the
types and hence it opens a space for stupid bugs when user e.g. passes
pointer to LIST where a node is expected or vice versa.

We fix this by making LIST a structure which _contains_ LIST_NODE
instead of the typedef.