Make CONTAINEROF safer by checking type.
We have replaced (ptr) with (1 ? (ptr) : &((type *)0)->member) which
(for correct uses of the macros) is eqivalent. However the use of the
operator ?: forces compiler to check that (ptr) has the same type as the
expect member of the enclosing structure.

If that precondition fails, any sane compiler should report a
compile-time error.